Fix scheduled tasks days field size
This field needs to be at least 31 bits long to store all days in a month.
Other fields are also set to 32 bits because unsigned int length can change
depending on the system.

+++ admin/Sources/Services/ogAdmServer/sources/schedule.c
@@ -49,7 +49,7 @@ static void og_parse_months(uint16_t months_mask, int months[])
 	}
 }
 
-static void og_parse_days(uint16_t days_mask, int days[])
+static void og_parse_days(uint32_t days_mask, int days[])
 {
 	int i, j = 0;
